# backend/openapi_snapshot.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "My Darrin API",
                 "version":  "0.1.0"
             },
    "paths":  {
...
                  "/api/v1/admin/users":  {
                                              "get":  {
                                                          "tags":  [
                                                                       "Admin",
                                                                       "Admin"
                                                                   ],
                                                          "summary":  "Get All Users",
                                                          "operationId":  "get_all_users_api_v1_admin_users_get",
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "Successful Response",
                                                                                        "content":  {
                                                                                                        "application/json":  {
                                                                                                                                 "schema":  {

                                                                                                                                            }
                                                                                                                             }
                                                                                                    }
                                                                                    }
                                                                        },
                                                          "security":  [
                                                                           {
                                                                               "HTTPBearer":  [

                                                                                              ]
                                                                           }
                                                                       ]
                                                      }
                                          },
                  "/api/v1/admin/users/pending":  {
                                                      "get":  {
                                                                  "tags":  [
                                                                               "Admin",
                                                                               "Admin"
                                                                           ],
                                                                  "summary":  "Get Pending Users",
                                                                  "operationId":  "get_pending_users_api_v1_admin_users_pending_get",
                                                                  "responses":  {
                                                                                    "200":  {
                                                                                                "description":  "Successful Response",
                                                                                                "content":  {
                                                                                                                "application/json":  {
                                                                                                                                         "schema":  {

                                                                                                                                                    }
                                                                                                                                     }
                                                                                                            }
                                                                                            }
                                                                                },
                                                                  "security":  [
                                                                                   {
                                                                                       "HTTPBearer":  [

                                                                                                      ]
                                                                                   }
                                                                               ]
                                                              }
                                                  },
                  "/api/v1/admin/users/{user_id}":  {
                                                        "get":  {
                                                                    "tags":  [
                                                                                 "Admin",
                                                                                 "Admin"
                                                                             ],
                                                                    "summary":  "Get User",
                                                                    "operationId":  "get_user_api_v1_admin_users__user_id__get",
                                                                    "security":  [
                                                                                     {
                                                                                         "HTTPBearer":  [

                                                                                                        ]
                                                                                     }
                                                                                 ],
                                                                    "parameters":  [
                                                                                       {
                                                                                           "name":  "user_id",
                                                                                           "in":  "path",
                                                                                           "required":  true,
                                                                                           "schema":  {
                                                                                                          "type":  "integer",
                                                                                                          "title":  "User Id"
                                                                                                      }
                                                                                       }
                                                                                   ],
                                                                    "responses":  {
                                                                                      "200":  {
                                                                                                  "description":  "Successful Response",
                                                                                                  "content":  {
                                                                                                                  "application/json":  {
                                                                                                                                           "schema":  {

                                                                                                                                                      }
                                                                                                                                       }
                                                                                                              }
                                                                                              },
                                                                                      "422":  {
                                                                                                  "description":  "Validation Error",
                                                                                                  "content":  {
                                                                                                                  "application/json":  {
                                                                                                                                           "schema":  {
                                                                                                                                                          "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                      }
                                                                                                                                       }
                                                                                                              }
                                                                                              }
                                                                                  }
                                                                }
                                                    },
                  "/api/v1/admin/users/{user_id}/activate":  {
                                                                 "put":  {
                                                                             "tags":  [
                                                                                          "Admin",
                                                                                          "Admin"
                                                                                      ],
                                                                             "summary":  "Activate User",
                                                                             "operationId":  "activate_user_api_v1_admin_users__user_id__activate_put",
                                                                             "security":  [
                                                                                              {
                                                                                                  "HTTPBearer":  [

                                                                                                                 ]
                                                                                              }
                                                                                          ],
                                                                             "parameters":  [
                                                                                                {
                                                                                                    "name":  "user_id",
                                                                                                    "in":  "path",
                                                                                                    "required":  true,
                                                                                                    "schema":  {
                                                                                                                   "type":  "integer",
                                                                                                                   "title":  "User Id"
                                                                                                               }
                                                                                                }
                                                                                            ],
                                                                             "responses":  {
                                                                                               "200":  {
                                                                                                           "description":  "Successful Response",
                                                                                                           "content":  {
                                                                                                                           "application/json":  {
                                                                                                                                                    "schema":  {

                                                                                                                                                               }
                                                                                                                                                }
                                                                                                                       }
                                                                                                       },
                                                                                               "422":  {
                                                                                                           "description":  "Validation Error",
                                                                                                           "content":  {
                                                                                                                           "application/json":  {
                                                                                                                                                    "schema":  {
                                                                                                                                                                   "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                               }
                                                                                                                                                }
                                                                                                                       }
                                                                                                       }
                                                                                           }
                                                                         }
                                                             },
                  "/api/v1/admin/users/{user_id}/deactivate":  {
                                                                   "put":  {
                                                                               "tags":  [
                                                                                            "Admin",
                                                                                            "Admin"
                                                                                        ],
                                                                               "summary":  "Deactivate User",
                                                                               "operationId":  "deactivate_user_api_v1_admin_users__user_id__deactivate_put",
                                                                               "security":  [
                                                                                                {
                                                                                                    "HTTPBearer":  [

                                                                                                                   ]
                                                                                                }
                                                                                            ],
                                                                               "parameters":  [
                                                                                                  {
                                                                                                      "name":  "user_id",
                                                                                                      "in":  "path",
                                                                                                      "required":  true,
                                                                                                      "schema":  {
                                                                                                                     "type":  "integer",
                                                                                                                     "title":  "User Id"
                                                                                                                 }
                                                                                                  }
                                                                                              ],
                                                                               "responses":  {
                                                                                                 "200":  {
                                                                                                             "description":  "Successful Response",
                                                                                                             "content":  {
                                                                                                                             "application/json":  {
                                                                                                                                                      "schema":  {

                                                                                                                                                                 }
                                                                                                                                                  }
                                                                                                                         }
                                                                                                         },
                                                                                                 "422":  {
                                                                                                             "description":  "Validation Error",
                                                                                                             "content":  {
                                                                                                                             "application/json":  {
                                                                                                                                                      "schema":  {
                                                                                                                                                                     "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                                 }
                                                                                                                                                  }
                                                                                                                         }
                                                                                                         }
                                                                                             }
                                                                           }
                                                               },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                       "ValidationError":  {
                                                               "properties":  {
                                                                                  "loc":  {
                                                                                              "items":  {
                                                                                                            "anyOf":  [
                                                                                                                          {
                                                                                                                              "type":  "string"
                                                                                                                          },
                                                                                                                          {
                                                                                                                              "type":  "integer"
                                                                                                                          }
                                                                                                                      ]
                                                                                                        },
                                                                                              "type":  "array",
                                                                                              "title":  "Location"
                                                                                          },
                                                                                  "msg":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Message"
                                                                                          },
                                                                                  "type":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Error Type"
                                                                                           },
                                                                                  "input":  {
                                                                                                "title":  "Input"
                                                                                            },
                                                                                  "ctx":  {
                                                                                              "type":  "object",
                                                                                              "title":  "Context"
                                                                                          }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "loc",
                                                                                "msg",
                                                                                "type"
                                                                            ],
                                                               "title":  "ValidationError"
                                                           }
                                   },
                       "securitySchemes":  {
                                               "HTTPBearer":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer"
                                                              }
                                           }
                   }
}